setup painter plugins in the "painters" directory
* for a field type "FieldType" (aka HaloFile, TPMSnapshot),
there must be a directory in the "painters" dir with the
name "FieldTypePainter". This directory must hold a file
"__init__.py" that will be imported. The name of the painter
class defined in "__init__.py" must be the name of
the parent directory, i.e., "FieldTypePainter"

* also setup optional arguments in the Painter classes
to use "&" as the option marker

* can read list of arguments/options from file by
using "@" prefix before file name