adding meta data I/O to the power 1D case; should be backwards compatible
* Power1DStorage now takes an `edges` keyword and other keywords
are treated as metadata. This information is written to file
on lines that are "commented", i.e., they begin with `#`. **This allows
us to continue to use numpy.loadtxt (with the default comments=#) to
read only the data columns portion of the file, as was done before**
* Added files.ReadPower1DPlainText that reads new format of 1D output
files. Returns data, metadata where metadata includes `edges` and
any other meta info