restricting functionality of re-binning in order to avoid errors
so it turns out that you can't just re-bin to arbitrary bins without
introducing inaccuracies, i.e., you can't recover the proper information
if the bins are overlapping. so the simplest solution is to restrict to the
case where the new shape is an integer factor, such that we
can combine N bins to make the new array. if extra bins are
present, those are removed from the end