Add a read method to the painter plugin
We shall start migrating the painters to this infrastructure.
After the migration is done, we can just rename the painters to Readers
and move Painter.paint to power.py, or some seperate place (e.g. if it
shall be reused by a paint.py that paints points to a mesh)