distinguish weigth and mass in readers.
Because they are different.

power.py may need a flag --massweighted to use mass for weight.

This is to pave out the way to re-structure the .read interface.
I think the iterator iterface is broken by nature because the order of
reading is implicit. In several applications we demand a defined
order of reading, e.g. FOF reads ID and Position seperately -- this should
be explicitly visible in the source code, and currently it is unclear.