Refactor measurepower and power.py
The interface and logic in measurepower() is hugely simplified.
It does no more than project the 3D power spectrum to a prefered basis,
be it 1d, (k,mu), or (k, poles)

This commit paves the way to reuse the code in measurepower.py to measure
correlation function, in 1d, (k,mu) or (k, poles).