Avoid preparser altogether.
It is not nessesary.

For YAML configuration files, the plugins are loaded with X before strings
are converted to Plugin instances. (if this is not the case, then the
YAML file shall be arranged such that the plugin is decleared before the usage)

For direct commandline, as long as -X occurs
before the plugin name is referred, parsing won't fail.

For indirect commandline, the same as above.

This commit would have break power-parallel.py and corr-parallel.py

They insist passing in args fro the config.param_file to the parser.
It appears to be unnesseary, since the schecular will pass in the
parameter files anyways.