Allow decorator declaration of columns.
While implementing LPTParticles I realized columns may depend on
variables that are only available after __init__. (set_redshift)

In this case
we'd better allow declaring them as functions rather than filling
a dictionary during __init__.

The code looks cleaner with this approach -- we no longer need
to maintain the consistency bettwen hardcolumns and get_hardcolumn.
The decorator marks the function and it is automatically collected.

This won't work if the list of hardcolumns depends on the instance
(Source.File and Source.Array). They still override the full functions.