Avoid looking at test data in build/ directory.
This was the reason Lukas was seeing nested build directories.
The old logic thinks files in the build/ directory are also test
data files, and installs them into a nested directory. --
this goes on and on and on till the machine spends more time looking
for test files and may evetually crash the file system (too many nested
levels).