Simplify casting logic in FFTPower.
FFTPower now hosts two properties, first and second, both
are asserted to be MeshSource objects.

The resampling occurs in paint. So if a CatalogMesh object has the
wrong resolution, it will first be painted to a wrong
resolution then resampled. We may be able to
tweak the paint method of a CatalogMesh object to avoid the resampling
but I don't think it is desirable; for the compensation and perhaps
other 'actions' may depend on the resolution of the intrinsic Nmesh.