Add a value to to_mesh.
There are two types of weights.

The original 'weight' stands for the effective number of particles.
The second weight, 'Value' stands for the field that we are averaging.

This is crucial for dealing with multiple particle species (and masses).
We revealed the issue when Biwei was looking at hydro simulations with nbodykit.