add optimized selection function for dask arrays with test
This builds on the idea that the size of catalog is fixed after a certain point in the dask task graph.
So when we perform a selection, we can insert the selection task directly after the point in the
graph when the size last changed. The subsequent operations are only performed on the
selected data.

This saves a good amount of time -- for example, computing "Position" on the randoms catalog
if slicing only to a single redshift bin