Add stricter tests for FOF.
I noticed the output of FastPM fof catalog shows some variation
going from 1 to 4 ranks. The difference is concentrated near the
edges or domains, and it can affect even relatively large halos
to some extent. (though I was looking at a 5 step simulation.)

I still do not know what is causing the difference.

I checked that the position of particles are 1e-6 identical.

These FOF tests asserts the behavior of the parallel merging is
correct with simple cases that goes cross the domain boundary.