bigfile shall look for header starting from root.
As specified in the documentation. The rational is that
this way one can always easily specify a special header block in not
in the dataset. If it were relative to the dataset then there is
no way to load the header if it is outside.

This fix solves a few occasions where scripts mysteriously stopped
working when loading catalog.