when summing integers, use numpy.sum.
On some occasions when code is ran in e.g. a notebook,
numpy.sum is imported to replace sum, and the behaviro of
numpy.sum([]) is to 0.0 rather than builtin.sum([]) which returns 0.

numpy.sum([]) returning 0.0 causes downstream code expecting integers to fail,
thus copying the code out of nbodykit won't work.

We shall not serve as an poor example in this context.