Add a take method to BinnedStatistic.
The take method behaves more or less like the traditional numpy indexing.
Either a boolean mask of the same shape of data, or boolean masks along
each axis is accept. It is more powerful and sometimes allows more
intuitive code segments than the current sel method; see the example in
docstring.

The full boolean mask is projected along each axis then 'and'ed to the
given mask along the axis.

If a list of indices is given along any axis, it is first converted to
a mask.

In the end the booleans are converted to lists because the finalize method
takes only indices.