Remove Catalog interface from CatalogMesh.
This is required to add a compute method to MeshSource. Because
if CatalogMesh is a CatalogSource then CatalogSource.compute comes
into the way.

I think this PR also sheds some light on how to deal with the temporary
variables created during FKP/convpower. They probably shall be saved
as attributes of the CatalogMesh objects directly, rather than stored
as columns of the underlying sources.