Support dk=0 and dr=0 in FFTPower and FFTCorr
This PR also changes the file format of FFTPower.save and FFTCorr.save,
as they now also need to save the centers, which are no longer strictly
derived from the edges.

BinnedStatistic is extended to accept bin label coordinates as an external
argument. Of course rebinning would totally mess up the centers.

ProjectedFFTPower is left unchanged because it may eventually be rewritten with
the new 'transpose interface' of pmesh and improved support on 2d / 1d.

The edge property on FFTPower and FFTCorr, attached by the run method are now
removed. Use the edge property of the power object and corr object instead.