Remove the use of CurrentMPIComm.set in tests
Using CurrentMPIComm.set has a lot of problems. It doesn't restore the
previous default communicator, thus causing tests that does not call
CurrentMPIComm.set to fail in unpredictable ways.

We shall instead rely on explicitly passing in the communicator in the
low level tests.

CurrentMPIComm.enter() is tested in the high level tests.