Throttle during painting.
If some ranks receive too many partilces, throttle the chunksize.
This fixed random OOM errors, which may be the true cause of #525.

The current solution is not ideal. It is possible that the number
of particles on some ranks will not decrease as we decrease chunksize.

We can look into fixing that when it becomes a real problem.