mocks from galaxy power spectrum for a non-gaussian field (#662)
Hi, I am working on a project based on non-gaussian fields. So while generating mocks using nbodykit (let's say lognormal mocks), one requires to give the matter power spectrum(as a functional argument ) and bias (as a number) as input (among other parameters). This is fine as long as we are working are in a gaussian field but when it comes to non-gaussian field, bias is no longer an independent number but a function of k, hence a functional argument. I was wondering if this feature could be added where one could generate mocks in non-gaussian fields. I have some working code that I could share if you are interested. :)
Thanks,